scan engine subregion frame scan
1) ImageDisplay.h:
ROIActions needs ROI name instead of index

2)LaserScanning.c:
Took out mechanical shutter VChan and instead changed the
VChanROIStimulate to VChanROIShutter.  Rectangle ROIs can be drawn on
the scanned image and restored together with the scan settings back to
the scan engine. Still need to implement function for scan engine to
select which subregion ROI to scan.