project switchpoint from CVI 2013 to CVI 2015 and NI Vision 2014 to NI Vision 2015
Came a cross a bug in the NI Image display, when scanning and updating
the display and passing the mouse over the display, the display function
would freeze causing the imgDisplayTSV lock in the ImageBuilder in
LaserScanning.c to be obtained but not released.

Cannot figure out the cause of this error, thus upgrading to 2015
version will be attempted.