scan engine and pockells cell fixes
1) LaserScanning.c:
- ROI shutter for frame scans almost works, there is a delay though and
this makes the edges of the image to be jagged when the pockells cell is
turning off the light which should happen when the mirror turns around.

- There are problems with NI Display of images while scanning and
howering the mouse over the updated image can cause the image to freeze
and the software must be restarted.

- Frame Scan ROIs are not managed well, this makes it hard to switch
between various ROIs.

2) Pockells.c:
- Seems to work fine now. It can be also used to modulate the Pockells
cell when the fast axis scan mirror is turning around. The shuttering
delay for the signal coming from the scan engine must be still adjusted
carefully.

Next: will try to fix the frame scan ROI management.